requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
mediapipe>=0.10.0 